requirements.txt
tensorflow
numpy
opencv-python
streamlit
scikit-learn
Pillow